backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
pydantic-settings
sqlalchemy
psycopg
alembic
python-multipart
pillow
pytesseract
openai
yt-dlp
requests
openai-whisper
ffmpeg-python